plugins/browser: use size_t to store the return value of find_first_of()
esteidAPI.cpp:423: warning: comparison between signed and unsigned integer expressions

git-svn-id: https://esteid.googlecode.com/svn/plugins/browser/trunk@2773 0d7e1ef0-d974-fc83-f0b3-ccc071561e0e

+++ esteidAPI.cpp
@@ -420,7 +420,7 @@ std::vector<std::string> esteidAPI::stringSplit(std::string str, std::string sep
 {
     std::vector<std::string> results;
 
-    for (int found = 0; found != std::string::npos; found = str.find_first_of(separator)) {
+    for (size_t found = 0; found != std::string::npos; found = str.find_first_of(separator)) {
         if (found > 0) {
             results.push_back(str.substr(0, found));
             str = str.substr(found+1);
